+++ src/maximumflow.jl
@@ -11,12 +11,12 @@ export MaximumFlowResult
 
 
 struct MaximumFlowResult
-    path::Array{Connection,1}
+    path::Vector{Connection}
     flow::Float64
 end
 
 struct MaximumFlowProblem
-    connections::Array{Connection,1}
+    connections::Vector{Connection}
 end
 
 
@@ -69,7 +69,7 @@ function solve(problem::MaximumFlowProblem)
 
     cns = problem.connections
 
-    function leftexpressions(node::Int64, nodes::Array{Connection,1}, model)
+    function leftexpressions(node::Int64, nodes::Vector{Connection}, model)
         lst = []
         for conn in nodes
             if conn.to == node
@@ -86,7 +86,7 @@ function solve(problem::MaximumFlowProblem)
         return expr
     end
 
-    function rightexpressions(node::Int64, nodes::Array{Connection,1}, model)
+    function rightexpressions(node::Int64, nodes::Vector{Connection}, model)
         lst = []
         for conn in nodes
             if conn.from == node